+++ LSPI/sample.h
@@ -5,7 +5,7 @@ struct sample
 	float angle;
 	float angular_velocity;
 	int action;
-	float reward;
+	int reward;
 	float final_angle;
 	float final_angular_velocity;
 	int terminal;
